Move Util/fillUndefinedWithClosest → CoreTableUtils/interpolateRowValuesWithTolerance
