+++ grapher/utils/Util.test.ts
@@ -29,12 +29,10 @@ import {
     findClosestTimeIndex,
     trimArray,
     getDropIndexes,
-    fillUndefinedWithClosest,
 } from "grapher/utils/Util"
 import { strToQueryParams } from "utils/client/url"
 import { SortOrder } from "coreTable/CoreTableConstants"
 import { ScaleType } from "grapher/core/GrapherConstants"
-import { InvalidCellTypes } from "coreTable/InvalidCells"
 
 describe(findClosestTime, () => {
     describe("without tolerance", () => {
@@ -461,69 +459,5 @@ describe(sortNumeric, () => {
                 (o) => o.a
             )
         ).toEqual([{ a: 1 }, { a: 2 }, { a: 3 }, { a: 3 }, { a: 4 }, { a: 8 }])
-    })
-})
-
-describe(fillUndefinedWithClosest, () => {
-    it("handles empty array", () => {
-        expect(fillUndefinedWithClosest([], "value", "time", 2)).toEqual([])
-    })
-    it("leaves array unchanged if tolerance = 0", () => {
-        const result = fillUndefinedWithClosest(
-            [
-                { value: 1, time: 0 },
-                { value: undefined, time: 1 },
-                { value: undefined, time: 2 },
-                { value: 3, time: 3 },
-            ],
-            "value",
-            "time",
-            0
-        )
-        expect(result[1].value).toEqual(InvalidCellTypes.NoValueWithinTolerance)
-        expect(result[2].value).toEqual(InvalidCellTypes.NoValueWithinTolerance)
-    })
-    it("fills in gaps in simple case", () => {
-        const result = fillUndefinedWithClosest(
-            [
-                { value: 1, time: 0 },
-                { value: undefined, time: 1 },
-                { value: undefined, time: 2 },
-                { value: 3, time: 3 },
-            ],
-            "value",
-            "time",
-            2
-        )
-        expect(result.map((r) => r.value)).toEqual([1, 1, 3, 3])
-        expect(result.map((r) => r.time)).toEqual([0, 0, 3, 3])
-    })
-    it("fills in initial and trailing values", () => {
-        const result = fillUndefinedWithClosest(
-            [
-                { value: undefined, time: 0 },
-                { value: InvalidCellTypes.NaNButShouldBeNumber, time: 1 },
-                { value: 1, time: 2 },
-                { value: InvalidCellTypes.UndefinedButShouldBeNumber, time: 3 },
-                { value: undefined, time: 4 },
-                { value: undefined, time: 5 },
-                { value: 3, time: 6 },
-                { value: undefined, time: 7 },
-            ],
-            "value",
-            "time",
-            1
-        )
-        expect(result.map((r) => r.value)).toEqual([
-            InvalidCellTypes.NoValueWithinTolerance,
-            1,
-            1,
-            1,
-            InvalidCellTypes.NoValueWithinTolerance,
-            3,
-            3,
-            3,
-        ])
-        expect(result.map((r) => r.time)).toEqual([0, 2, 2, 2, 4, 6, 6, 6])
     })
 })
